for:each directive to Render List
When using the for:each directive, use for:item="currentItem" to access the current item. This example doesn’t use it, but to access the current item’s index, use  for:index="index".

To assign a key to the first element in the nested template, use the key={uniqueId} directive.