iterator Directives
To apply a special behavior to the first or last item in a list, use the iterator directive,  iterator:iteratorName={array}. Use the iterator directive on a template tag.

Use iteratorName to access these properties:

value—The value of the item in the list. Use this property to access the properties of the array. For example, iteratorName.value.propertyName.
index—The index of the item in the list.
first—A boolean value indicating whether this item is the first item in the list.
last—A boolean value indicating whether this item is the last item in the list.