LDS Editing a Record
To edit a record, pass the ID of the record and the corresponding object API name to be edited. Specify the fields using the fields attribute, or layout-type attribute to load all the fields defined on the Full or Compact layout.

When record-id is passed, edit mode loads the form with input fields displaying the specified record's field values. The form also displays Cancel and Save buttons.

This example creates an editable two-column form for an account record using the compact layout. Place the form on an account record page to inherit its record-id and object-api-name properties. The onsubmit attribute specifies an action to override the handler for the submit.